// ejercicios/c13-docker/frontend/src/schemas/libroSchema.ts
import { z } from 'zod'

export const libroSchema = z.object({
  titulo: z.string().trim().min(1, 'El título es obligatorio'),
  autor: z.string().trim().min(1, 'El autor es obligatorio'),
  precio: z.coerce.number().positive('El precio debe ser mayor a 0'),
  genero: z.string().min(1, 'El género es obligatorio'),
  disponible: z.boolean(),
})

export type LibroFormData = z.infer<typeof libroSchema>